+++ packages/helper/src/extension.ts
@@ -1,5 +1,9 @@
 import * as vscode from "vscode";
-import { Subscriptions, HoverProvider } from "./features";
+import {
+  NimbusDependenciesChecker,
+  Subscriptions,
+  HoverProvider,
+} from "./features";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -7,9 +11,10 @@ export function activate(context: vscode.ExtensionContext) {
   // Use the console to output diagnostic information (console.log) and errors (console.error)
   // This lin``e of code will only be executed once when your extension is activated
   console.log(
-    'Congratulations, your extension "@nimbus-ds/vscode-components" is now active!'
+    'Congratulations, your extension "@nimbus-ds/helper" is now active!'
   );
 
+  const nimbusDependenciesChecker = new NimbusDependenciesChecker();
   const subscriptions = new Subscriptions();
   const hoverProvider = new HoverProvider();
 
@@ -18,6 +23,14 @@ export function activate(context: vscode.ExtensionContext) {
     hoverProvider
   );
 
+  nimbusDependenciesChecker.check();
+
+  const disposable = vscode.commands.registerCommand(
+    "nimbus-helper.checknimbusdependencies",
+    () => nimbusDependenciesChecker.check()
+  );
+
   subscriptions.setSubscriptions(hover);
+  subscriptions.setSubscriptions(disposable);
   context.subscriptions.push(...subscriptions.getSubscriptions());
 }